linux-xlnx: Add e3xx support for 2017.3 release
This change ports patching and configuration required to build
e3xx boards against linux-xlnx 2017.3 from the meta-xilinx layer.

This was a rather involved change that included porting the device
tree sources to work with the 4.9 Kernel, and a key change here
was to leverage the zynq-7000.dtsi instead of using generated dts
files.

There was one non-e3xx specific patch that was backported from a
newer xilinx kernel ("char: xilinx_devcfg: Fix regression") that
was required to keep the firmware loader from timing out. When
these patches are ported to the next Kernel from meta-xilinx, this
patch can probably be removed.